Repair delete image in admin section

+++ src/main.ts
@@ -174,7 +174,13 @@ export const app: App = {
                     });
             }
 
-            uploadForm.style.display = 'block';
+            const currentURL = window.location.href;
+            //console.log("Aktualny adres URL strony:", currentURL);
+            if (currentURL.includes('/#/loginForm')) {
+                uploadForm.classList.toggle('active');
+            } else {
+                console.log("false", currentURL);
+            }
         });
 
     },
@@ -197,7 +203,7 @@ export const app: App = {
 
                     console.log('data', data)
                     const imagesHTML = data.map((image: { value: any; }) => `
-                        <img class="border-2 border-navy-strongBlue rounded-xl first:mt-0 mt-5 lg:mt-8" src="/uploads/${image}" alt="">
+                        <img name="${image}" class="border-2 border-navy-strongBlue rounded-xl first:mt-0 mt-5 lg:mt-8" src="/uploads/${image}" alt="">
                     `).join('');
 
 
@@ -239,7 +245,7 @@ export const app: App = {
                         ${index + 1}
                         <img class="w-6 h-6" src="/uploads/${item}" alt="">
                         <p class="">${item}</p>
-                        <button id="removeButton" name="${item}" type="button" class="flex">
+                        <button name="${item}" type="button" class="remove flex">
                             Usuń
                             <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6 ml-3">
                                 <path stroke-linecap="round" stroke-linejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" />
@@ -286,14 +292,24 @@ export const app: App = {
             });
         
         }
-        const chosenImageToDelate = document.getElementById('removeButton');
-        // Przykład wywołania funkcji deleteImage dla konkretnego pliku
-        //const fileNameToDelete = chosenImageToDelate. // Tutaj podaj nazwę pliku, który chcesz usunąć
-        
-        console.log('chosen', chosenImageToDelate);
 
-        
-        deleteImage(chosenImageToDelate);
+        const pictureList = document.getElementById('pictureList'); 
+        pictureList?.addEventListener('click', (event) => {
+            console.log('pictureList', event);
+            if((event.target as HTMLElement).classList.contains('remove')) {
+                const chosenImageToDelate = (event.target as HTMLButtonElement).getAttribute('name');
+                if(chosenImageToDelate && confirm('Czy na pewno chcesz usunąć to zdjęcie?')) {
+                    deleteImage(chosenImageToDelate);
+                    if(event.target) {
+                        // @ts-ignore we know this element exists
+                        event.target.parentNode.remove();
+                    }
+                    const image = document.querySelector('img[name="' + chosenImageToDelate + '"]');
+                    image!.remove();
+                }
+            }
+        });
+     
     }
 
 };
